Fix SBT build
In the previous PR to migrate to CircleCI, build.sbt was simplified in
order to properly cross compile to all specified Scala versions.
However, as it turns out this caused nothing to actually get compiled.
This reverts said changes so that this library is properly cross
compiled and published.
Note that since v10.1.x `akka-http` requires `Sakka-stream`.